First running version with multiple NLTE elements
This version allows computation of spectra with more than one element in NLTE.
A control file has been added (SPECIES_LTE_NLTE.dat) that contains flags and ressource files for each element of interest.
Its path is indicated in the launch script by  NLTEINFOFILE. This replaces the MODELATOMFILE, DEPARTUREFILE and  DEPARTBINARY keywords in the script. There is no need to have the NLTE/LTE flag in the line list anymore.
Securities were added, in case a departure file is missing: b=1 departure coefficients are then used. If the model atom level identification is missing in the line list b=1 is used for these levels.
Not many warnings are issued. Check the log file carefully!

A few other things were fixed in this version, including H NLTE calculations.