small changes
character insensitive check of species from model atom in bsyn.f

added possibility to compute radius where tau_lambda=1 in bsynb.f

allowed computations to continue even if one wavelength gives negative result for intensity or flux. In that case the value is flagged as negative. Might result in NaN in spectrum. The use for this is to overcome difficult situation with very optically thick lines in very extended models.