Fixed crash with extra comments in model atom
Fixed crash when there are extra comments in the model atom file. For example, if there are comments after name of the atom, but before the abundance or number of lines (e.g. * ABUND      AWGT or * NK     NLINE   NCONT   NRFIX) it would crash. Now added check for header while loop to skip any comments in-between.