Simpler hlinop and no offset between core and wing of HI lines
this version reverses to hlinop.f as it is on GitHub from  Paul Barklem.
This was to test if there was a problem with the satellites as implemented by Regner Trampedach, but this was not the issue.
The issue (profiles with an offset in the wing) is with the approxiamte convolution of broadening mechanisms: in the core only the dominant was kept, and outside the core all mechanisms are added.
In this version, the core is defined with a zero width so that all mechanisms are added everywhere .
Note that this remains approximate, as it is not a real convolution.